Rework index build concurrency
We need to prevent checkpointer from processing trees, which build is
in-progress.  Currently we insert meta-information into sys trees and write
tree headers under oTablesMetaLock.  But this is hard to reproduce in recovery,
where we reconstruct DDL operations from sys trees changes, given it's not
acceptable to hold oTablesMetaLock during index build.

Now, we build indices in three phases.

 1. Update meta-information in system trees and insert shared root info
    placeholders for new trees (under oTablesMetaLock).
 2. Write the new trees data.
 3. Insert new tree headers and delete place holders (under oTablesMetaLock).

So, checkpointer can't reach the new tree until the meta-information is
complete.  Also checkpoint numer in tree headers is valid.  This logic is easy
to replay during recovery.