Handle case when page is failed to be compressed
In this case we still need to fit the on-disk page image to ORIOLEDB_BLCKSZ.
So, we need to fit both OCompressHeader and the page contents.  We do this by
skipping writing BTreePageHeader.o_header, which makes sense only for
in-memory page image.  Also, we skip BTreePageHeader.checkpointNum, because
it's already present in OCompressHeader.  For simplicity we rearrange fields
of BTreePageHeader, since we anyway have to break binary compatibility.
The relevant test case is added.