Fix: Use fake COMMIT_HASH if no .git history (Docker environment, building from tar.gz)
In the Docker environment, the .git subdirectory is removed to optimize the final image size,
making it impossible to retrieve the actual COMMIT_HASH.
This fix provides a temporary fake COMMIT_HASH
with a value of 0000000000000000000000000000000000000000 to handle this situation.

This fix is also useful for building from orioledb (zip) or (tar.gz) source code
when there is no .git information.

With this fix, the regression tests related to `orioledb_commit_hash()`
now pass successfully even without git commit hash information.