Split undo log into data undo log and system undo log
Logical decoding requires us to keep system catalog for not-yet-decoded
part of WAL.  For OrioleDB that means keeping undo log for systems trees for
that part of WAL.  That could bloat undo log size.  In order to avoid this,
we split undo log into regular and system.  Only system undo log we're going
to retain for logical decoding purposes.