Add test `test_checkpoint_autonomous_page_hikey`
Add new test `test_checkpoint_autonomous_page_hikey` to test if a page
which became autonomous was merged by checkpointer and if it is sparse.
It is necessary to make sure that a transaction and the checkpointer hit
same page.
Add new function `tbl_structure_to_json` to parse output of the
`orioledb_tbl_structure` function.