Add SKIP_INSTALL Handling in Makefile for Better Testing
This commit adds conditional logic to handle the SKIP_INSTALL flag in the Makefile.
When SKIP_INSTALL is enabled, the installation step is skipped, allowing tests to run as a non-root user even if the installation was performed as root.
This is especially useful in CI environments where the extension is preinstalled, focusing solely on functionality testing;
  make SKIP_INSTALL=1 USE_PGXS=1 installcheck
Additionally, it helps avoid unnecessary reinstallation when the extension is installed via package managers in the future (such as apt, apk, or yum).