OrioleDB database cluster rewind
Implements rewind of OrioleDB and Postgres tables to a point
in the past specified by one of the three alternative ways:
- to xid/oxid pair
- to timestamp with time zone
- by time in seconds back from now

Rewind for 'orioledb' tables is implemented using OrioleDB undo logs.

Rewind for 'heap' tables is implemented by changing visibility info and
delaying vacuum. Using the rewind mode on a workload that does significant
writes to heap tables can lead to accumulation of significant amount
of invisible past tuple versions in the table, that can't be cleaned by
VACUUM. It's not recommended.

Depth of rewind is limited by customer by the time and the number of
transactions (orioledb.rewind_max_time, orioledb.rewind_max_transactions).
It's also limited by an uptime of a database cluster: at the start of a
cluster all transactions retained for rewind are premanently fixed and
could not be rewound anymore.

After rewind, the database cluster is shut down. It's a responsibility of a
caller to start it after rewind.

Co-authored-by: Alexander Korotkov <akorotkov@postgresql.org>