Orioledb rewind
Implements rewind of Oriole and Postgres tables to a point
in the past specified as:
- xid/oxid pair
- timestamp
- time in seconds from now

Rewind for orioledb tables implemented via internal Oriole undo logs.

Rewind for heap tables is implemented by changing visibility info and
delaying vacuum. Use of the rewind mode on a workload that does significant
writes to heap tables leads to accumulation of significant amount of past
tuple versions in the table so it's not recommended.

Depth of rewind is limited by both orioledb.rewind_max_time and
orioledb.rewind_max_transactions GUC's.

After rewind database cluster is shut down. It's a responsibility of a caller
to start it after rewind.