Fix OIndex deserialize and serialize size calculation
It's a follow-up to 9b89ecf50958d605 that comments a bug introduced in efb3e2a58c6f1
but not fixes it to not break binary compatibility of databases.

This commit fixes a bug and breaks binary compatibility with databases previous to it.

Bump ORIOLEDB_BINARY_VERSION.

Per-report: Artur Zakirov