Fix for issue #231
This commit implements map from CommandId to the UndoLocation of the first
undo record for that command.  It is used to determine visibility within the
same transaction and to detect "self-updated" tuples.  That is a bit tricky,
assuming PostgreSQL can switch execution between commands.  However, that
could only happen for "subcommands," such as trigger execution.  However, the
execution of a command finishes after all of its subcommands, so a comparison
of undo positions should be fine for checking if the given change belongs to
some of the previous commands.

These changes allow the implementation of correct visibility of the same
transaction changes for cursors.  Additionally, this enables us to eliminate
the executor hook.