Support logical replication with REPLICA IDENTITY FULL
This mode is necessary for logical replication of UPADTE's and DELETE's in and OrioleDB table without explicit PRIMARY KEY.
Hidden ctid column that exists in this case is not visible and not logically replicated. Without these modes UPDATE's and DELETE's on a
replicated table are prohibited like in plain Postgres.

Syntax for setting replica identity is similar to PG, but other replica identity settings other than FULL and DEFAULT are not yet supported.