Added lockstep scan for arrays and fixed coercible calls
During Array scans like (SELECT ... WHERE id IN (1,3,4...)) orioledb
does full tree descend for each element of the array. This is
inefficient when the array is large and the elements are close to each
other in the index. This commit adds lockstep scan for arrays, which
descends the tree only once for all elements of the array and uses iterators.

Also fixed the problem with too many coercible calls during index scans (perf showed that
it was having significant impact on performance).
Even if the key was not coercible, orioledb was trying to check that again and
again during index scans.